Move ThreadFlags to embassy impl
